+++ R/format_path.R
@@ -12,11 +12,11 @@
 #'
 #' @examples
 #' format_path(c("file1", "file2"), "csv")
-#'
+#' 
 #' (dfs <- list_csv(tor_example("csv")))
-#'
+#' 
 #' format_path(names(dfs), "csv")
-#'
+#' 
 #' format_path(names(dfs), "csv", base = "home", prefix = "this-")
 format_path <- function(files, ext, base = ".", prefix = NULL) {
   if (missing(files)) abort("`files` can't be missing")