+++ tests/testthat/test-list_any.R
@@ -3,7 +3,8 @@ context("list_any")
 test_that("list_any with read.csv lists (file)named dataframes", {
   res <- list_any(
     tor_example("csv"),
-    utils::read.csv, regexp = "[.]csv$"
+    utils::read.csv,
+    regexp = "[.]csv$"
   )
 
   expect_is(res, "list")
@@ -14,7 +15,7 @@ test_that("list_any with read.csv lists (file)named dataframes", {
 test_that("list_any accepts lambda functions and formulas", {
   res <- list_any(
     tor_example("rdata"),
-    ~get(load(.x))
+    ~ get(load(.x))
   )
 
   expect_is(res, "list")
@@ -33,7 +34,8 @@ test_that("list_any reads specific files extention in a mixed directory", {
   expect_is(
     list_any(
       tor_example("mixed"),
-      utils::read.csv, regexp = "[.]csv$"
+      utils::read.csv,
+      regexp = "[.]csv$"
     ),
     "list"
   )
@@ -61,7 +63,7 @@ test_that("list_any passes arguments to the reader function via `...`", {
   expect_is(
     list_any(
       tor_example("csv"),
-      ~read.csv(., stringsAsFactors = FALSE)
+      ~ read.csv(., stringsAsFactors = FALSE)
     )[[2]]$y,
     "character"
   )
@@ -69,7 +71,7 @@ test_that("list_any passes arguments to the reader function via `...`", {
 
 test_that("list_any with emtpy path reads from working directory", {
   expect_is(
-    list_any( , read.csv, "[.]csv")[[1]],
+    list_any(, read.csv, "[.]csv")[[1]],
     "data.frame"
   )
 })