+++ R/list_any.R
@@ -6,33 +6,32 @@
 #'
 #' @return A list.
 #'
-#' @family general functions to import data
-#' @export
-#'
 #' @examples
 #' tor_example()
-#'
+#' 
 #' path <- tor_example("csv")
 #' dir(path)
-#'
+#' 
 #' list_any(path, read.csv)
-#'
-#' list_any(path, ~read.csv(.x, stringsAsFactors = FALSE))
-#'
+#' 
+#' list_any(path, ~ read.csv(.x, stringsAsFactors = FALSE))
+#' 
 #' (path_mixed <- tor_example("mixed"))
 #' dir(path_mixed)
-#'
+#' 
 #' list_any(
-#'   path_mixed, ~get(load(.x)),
+#'   path_mixed, ~ get(load(.x)),
 #'   regexp = "[.]csv$",
 #'   invert = TRUE
 #' )
-#'
+#' 
 #' list_any(
-#'   path_mixed, ~get(load(.x)),
+#'   path_mixed, ~ get(load(.x)),
 #'   "[.]Rdata$",
 #'   ignore.case = TRUE
 #' )
+#' @family general functions to import data
+#' @export
 list_any <- function(path = ".",
                      .f,
                      regexp = NULL,