+++ R/tor_example.R
@@ -6,7 +6,6 @@
 #' @param path Length-1 character vector. A path to an available directory.
 #'
 #' @return A list of available directories or a path to them.
-#' @export
 #'
 #' @examples
 #' tor_example()
@@ -14,6 +13,8 @@
 #' tor_example("csv")
 #'
 #' dir(tor_example("csv"))
+#' @family helpers
+#' @export
 tor_example <- function(path = NULL) {
   if (is.null(path)) {
     dir(system.file("extdata", package = "tor"))