+++ NEWS.md
@@ -1,4 +1,8 @@
-# tor 1.0.0.9000 unreleased
+# tor 1.0.9001 unreleased
+
+# Patch
+
+* Refactor: For clarity source now uses the more common function(x) syntax. Text editors highlight `function()` which makes code easier to read.
 
 # tor 1.0.0