+++ NEWS.md
@@ -1,6 +1,10 @@
 # tor 1.0.9001 unreleased
 
-# Patch
+## Minor
+
+* New `load_*()` variants load each file in a directory into an environment.
+
+## Patch
 
 * Refactor: For clarity source now uses the more common function(x) syntax. Text editors highlight `function()` which makes code easier to read.